refactor(#652): move coaching page to app router
