+++ frontend/src/app/(scoreboard)/coaching/page.tsx
@@ -1,0 +1,10 @@
+import { Suspense } from 'react';
+import CoachingPage from './CoachingPage';
+
+export default function Page() {
+    return (
+        <Suspense>
+            <CoachingPage />
+        </Suspense>
+    );
+}
