+++ frontend/src/profile/coach/CoachTab.tsx
@@ -9,8 +9,8 @@ import {
 } from '@mui/material';
 import { useCallback } from 'react';
 
-import { displayEvent } from '../../coaching/customers/CoachingList';
-import UpcomingSessions from '../../coaching/customers/UpcomingSessions';
+import { displayEvent } from '../../app/(scoreboard)/coaching/CoachingList';
+import UpcomingSessions from '../../app/(scoreboard)/coaching/UpcomingSessions';
 import { Event } from '../../database/event';
 import { User } from '../../database/user';
 import Bio from '../info/Bio';
